Updates testing environment. Moves compiler to outer directory.
