+++ .github/workflows/testing-backend.yml
@@ -10,7 +10,7 @@ on:
       - 'master'
 jobs:
   run-controller-tests:
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
     steps:
       # Checkout repository code
       - uses: actions/checkout@v2
