+++ .github/workflows/testing-frontend.yml
@@ -10,7 +10,7 @@ on:
       - 'master'
 jobs:
   run-frontend-tests:
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
